Refactor: Separate bot logic into bot.rs module
- Move bot-related functions (OCR, message handlers) to src/bot.rs
- Simplify main.rs to focus on initialization and dispatcher setup
- Improve code organization and maintainability
